+++ compiler-core/src/analyse.rs
@@ -76,10 +76,34 @@ impl Inferred<PatternConstructor> {
     }
 }
 
+/// How the compiler should treat target support.
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub enum TargetSupport {
+    /// Target support is enfored, meaning if a function is found to not have an implementation for
+    /// the current target then an error is emitted and compilation halts.
+    ///
+    /// This is used when compiling the root package, with the exception of when using
+    /// `gleam run --module $module` to run a module from a dependency package, in which case we do
+    /// not want to error as the root package code isn't going to be run.
     Enforced,
+    /// Target support is enfored, meaning if a function is found to not have an implementation for
+    /// the current target it will continue onwards and not generate any code for this function.
+    ///
+    /// This is used when compiling dependencies.
     NotEnforced,
+}
+
+impl TargetSupport {
+    /// Returns `true` if the target support is [`Enforced`].
+    ///
+    /// [`Enforced`]: TargetSupport::Enforced
+    #[must_use]
+    pub fn is_enforced(&self) -> bool {
+        match self {
+            Self::Enforced => true,
+            Self::NotEnforced => false,
+        }
+    }
 }
 
 // TODO: This takes too many arguments.
@@ -612,6 +636,7 @@ fn infer_function(
         return_type: (),
         implementations: _,
     } = f;
+    let target = environment.target;
     let preregistered_fn = environment
         .get_variable(&name)
         .expect("Could not find preregistered type for function");
@@ -622,8 +647,7 @@ fn infer_function(
         .expect("Preregistered type for fn was not a fn");
 
     // Find the external implementation for the current target, if one has been given.
-    let external =
-        target_function_implementation(environment.target, &external_erlang, &external_javascript);
+    let external = target_function_implementation(target, &external_erlang, &external_javascript);
     let (impl_module, impl_function) = implementation_names(external, module_name, &name);
 
     // The function must have at least one implementation somewhere.
@@ -664,6 +688,17 @@ fn infer_function(
 
     // Assert that the inferred type matches the type of any recursive call
     unify(preregistered_type, type_.clone()).map_err(|e| convert_unify_error(e, location))?;
+
+    // Ensure that the current target has an implementation for the function.
+    // This is done at the expression level while inferring the function body, but we do it again
+    // here as externally implemented functions may not have a Gleam body.
+    if public && environment.target_support.is_enforced() && !implementations.supports(target) {
+        return Err(Error::UnsupportedPublicFunctionTarget {
+            name: name.clone(),
+            target,
+            location,
+        });
+    }
 
     let variant = ValueConstructorVariant::ModuleFn {
         documentation: doc.clone(),