+++ compiler-core/src/type_/tests/expression.rs
@@ -34,7 +34,7 @@ pub fn pure_gleam_function() {
 pub fn pure_gleam_1() { 1 + 1 }
 pub fn pure_gleam_2() { pure_gleam_1() * 2 }
 "#,
-        vec![
+        [
             (
                 "pure_gleam_1",
                 Implementations {
@@ -64,7 +64,7 @@ pub fn erlang_only_1() -> Int
 
 pub fn erlang_only_2() { erlang_only_1() * 2 }
 "#,
-        vec![
+        [
             (
                 "erlang_only_1",
                 Implementations {
@@ -80,7 +80,7 @@ pub fn erlang_only_2() { erlang_only_1() * 2 }
                     uses_erlang_externals: true,
                     uses_javascript_externals: false,
                 }
-            ),
+            )
         ],
     );
 }
@@ -95,7 +95,7 @@ pub fn all_externals_1() -> Int
 
 pub fn all_externals_2() { all_externals_1() * 2 }
 "#,
-        vec![
+        [
             (
                 "all_externals_1",
                 Implementations {
@@ -130,7 +130,7 @@ pub fn pure_gleam() {
   javascript_external_and_pure_body() + erlang_external_and_pure_body()
 }
 "#,
-        vec![
+        [
             (
                 "erlang_external_and_pure_body",
                 Implementations {
@@ -164,14 +164,14 @@ pub fn erlang_external_with_javascript_body() {
     assert_targets!(
         r#"
 @external(javascript, "foo", "bar")
-pub fn javascript_only() -> Int
+fn javascript_only() -> Int
 
 @external(erlang, "foo", "bar")
 pub fn erlang_external_and_javascript_body() -> Int { javascript_only() }
 
 pub fn all_externals() -> Int { erlang_external_and_javascript_body() }
 "#,
-        vec![
+        [
             (
                 "all_externals",
                 Implementations {
@@ -195,7 +195,7 @@ pub fn all_externals() -> Int { erlang_external_and_javascript_body() }
                     uses_erlang_externals: false,
                     uses_javascript_externals: true,
                 }
-            ),
+            )
         ],
     );
 }
@@ -212,7 +212,7 @@ pub fn javascript_external_and_erlang_body() -> Int { erlang_only() }
 
 pub fn all_externals() -> Int { javascript_external_and_erlang_body() }
 "#,
-        vec![
+        [
             (
                 "all_externals",
                 Implementations {
@@ -236,7 +236,7 @@ pub fn all_externals() -> Int { javascript_external_and_erlang_body() }
                     uses_erlang_externals: true,
                     uses_javascript_externals: true,
                 }
-            ),
+            )
         ],
     );
 }
@@ -246,10 +246,10 @@ pub fn function_with_no_valid_implementations() {
     assert_module_error!(
         r#"
 @external(javascript, "foo", "bar")
-pub fn javascript_only() -> Int
+fn javascript_only() -> Int
         
 @external(erlang, "foo", "bar")
-pub fn erlang_only() -> Int
+fn erlang_only() -> Int
 
 pub fn main() {
     javascript_only()